requirements.txt
runpod
torch
diffusers
transformers
accelerate
pillow
opencv-python
